Moved demo files to a single folder
- demo folder renamed
- added readme
- changed file paths
- changed Utils/BaseDemo to make it identical to master branch
