+++ UD-Viz-Core/examples/Demo_DatAgora_PartDieu/Demo.html
@@ -4,7 +4,7 @@
         <title>UD-Viz-Core</title>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link rel="stylesheet" href="../../src/Utils/DataViewer/css/Demo.css">
+        <link rel="stylesheet" href="./Demo.css">
     </head>
     <body>
         <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.0/jquery.min.js"></script>
