+++ UD-Viz-Core/index.html
@@ -56,7 +56,7 @@
                     </p>
                 </li>
                 <li>
-                    <h3><a href='examples/DemoData/Demo.html'>UD-Viz Demo Projet "Végétalisation de la Part-Dieu"
+                    <h3><a href='examples/Demo_DatAgora_PartDieu/Demo.html'>UD-Viz Demo Projet "Végétalisation de la Part-Dieu"
                     </a></h3>
                     <p>The demonstration used to dislay all data used regarding the revegetation of the Part-Dieu neighbourhood
                     </p>
